+++ gsad/src/gsad_gmp.c
@@ -5326,7 +5326,8 @@ create_target_gmp (gvm_connection_t *connection, credentials_t *credentials,
   CHECK_VARIABLE_INVALID (target_ssh_credential, "Create Target");
   if (strcmp (target_ssh_credential, "--"))
     CHECK_VARIABLE_INVALID (port, "Create Target");
-  CHECK_VARIABLE_INVALID (target_ssh_elevate_credential, "Create Target");
+  if (params_given (params, "ssh_elevate_credential_id"))
+    CHECK_VARIABLE_INVALID (target_ssh_elevate_credential, "Create Target");
   CHECK_VARIABLE_INVALID (target_smb_credential, "Create Target");
   CHECK_VARIABLE_INVALID (target_esxi_credential, "Create Target");
   CHECK_VARIABLE_INVALID (target_snmp_credential, "Create Target");
@@ -5349,8 +5350,11 @@ create_target_gmp (gvm_connection_t *connection, credentials_t *credentials,
                                                  "<port>%s</port>"
                                                  "</ssh_credential>",
                                                  target_ssh_credential, port);
-      ssh_elevate_credentials_element = g_strdup_printf (
-        "<ssh_elevate_credential id=\"%s\"/>", target_ssh_elevate_credential);
+      if (target_ssh_elevate_credential)
+        ssh_elevate_credentials_element = g_strdup_printf (
+          "<ssh_elevate_credential id=\"%s\"/>", target_ssh_elevate_credential);
+      else
+        ssh_elevate_credentials_element = NULL;
     }
 
   if (strcmp (target_smb_credential, "0") == 0)
@@ -5401,13 +5405,14 @@ create_target_gmp (gvm_connection_t *connection, credentials_t *credentials,
     reverse_lookup_unify ? reverse_lookup_unify : "0", port_list_id,
     alive_tests, allow_simultaneous_ips ? allow_simultaneous_ips : "1");
 
-  command = g_strdup_printf ("<create_target>"
-                             "%s%s%s%s%s%s%s%s"
-                             "</create_target>",
-                             xml->str, comment_element, ssh_credentials_element,
-                             ssh_elevate_credentials_element,
-                             smb_credentials_element, esxi_credentials_element,
-                             snmp_credentials_element, asset_hosts_element);
+  command = g_strdup_printf (
+    "<create_target>"
+    "%s%s%s%s%s%s%s%s"
+    "</create_target>",
+    xml->str, comment_element, ssh_credentials_element,
+    ssh_elevate_credentials_element ? ssh_elevate_credentials_element : "",
+    smb_credentials_element, esxi_credentials_element, snmp_credentials_element,
+    asset_hosts_element);
 
   g_string_free (xml, TRUE);
   g_free (comment_element);
@@ -6296,7 +6301,8 @@ save_target_gmp (gvm_connection_t *connection, credentials_t *credentials,
       && strcmp (target_ssh_credential, "0"))
     {
       CHECK_VARIABLE_INVALID (port, "Save Target");
-      CHECK_VARIABLE_INVALID (target_ssh_elevate_credential, "Save Target");
+      if (params_given (params, "ssh_elevate_credential_id"))
+        CHECK_VARIABLE_INVALID (target_ssh_elevate_credential, "Save Target");
     }
 
   if (str_equal (target_source, "manual"))
@@ -6328,8 +6334,12 @@ save_target_gmp (gvm_connection_t *connection, credentials_t *credentials,
                                                    "<port>%s</port>"
                                                    "</ssh_credential>",
                                                    target_ssh_credential, port);
-        ssh_elevate_credentials_element = g_strdup_printf (
-          "<ssh_elevate_credential id=\"%s\"/>", target_ssh_elevate_credential);
+        if (target_ssh_elevate_credential)
+          ssh_elevate_credentials_element =
+            g_strdup_printf ("<ssh_elevate_credential id=\"%s\"/>",
+                             target_ssh_elevate_credential);
+        else
+          ssh_elevate_credentials_element = NULL;
       }
 
     if (str_equal (target_smb_credential, "--"))
@@ -6371,13 +6381,14 @@ save_target_gmp (gvm_connection_t *connection, credentials_t *credentials,
       reverse_lookup_unify ? reverse_lookup_unify : "0", port_list_id,
       alive_tests, allow_simultaneous_ips ? allow_simultaneous_ips : "1");
 
-    g_string_append_printf (command,
-                            "%s%s%s%s%s%s"
-                            "</modify_target>",
-                            comment_element, ssh_credentials_element,
-                            ssh_elevate_credentials_element,
-                            smb_credentials_element, esxi_credentials_element,
-                            snmp_credentials_element);
+    g_string_append_printf (
+      command,
+      "%s%s%s%s%s%s"
+      "</modify_target>",
+      comment_element, ssh_credentials_element,
+      ssh_elevate_credentials_element ? ssh_elevate_credentials_element : "",
+      smb_credentials_element, esxi_credentials_element,
+      snmp_credentials_element);
 
     g_free (comment_element);
     g_free (ssh_credentials_element);
@@ -9978,27 +9989,27 @@ get_system_reports_gmp (gvm_connection_t *connection,
 
   if (strcmp (range_type, "duration") == 0)
     {
-      struct tm *time_broken;
-      time_broken = localtime (&now);
-      end_time.tm_year = time_broken->tm_year;
-      end_time.tm_mon = time_broken->tm_mon;
-      end_time.tm_mday = time_broken->tm_mday;
-      end_time.tm_hour = time_broken->tm_hour;
-      end_time.tm_min = time_broken->tm_min;
+      struct tm time_broken;
+      localtime_r (&now, &time_broken);
+      end_time.tm_year = time_broken.tm_year;
+      end_time.tm_mon = time_broken.tm_mon;
+      end_time.tm_mday = time_broken.tm_mday;
+      end_time.tm_hour = time_broken.tm_hour;
+      end_time.tm_min = time_broken.tm_min;
 
-      time_broken = localtime (&duration_start);
-      start_time.tm_year = time_broken->tm_year;
-      start_time.tm_mon = time_broken->tm_mon;
-      start_time.tm_mday = time_broken->tm_mday;
-      start_time.tm_hour = time_broken->tm_hour;
-      start_time.tm_min = time_broken->tm_min;
+      localtime_r (&duration_start, &time_broken);
+      start_time.tm_year = time_broken.tm_year;
+      start_time.tm_mon = time_broken.tm_mon;
+      start_time.tm_mday = time_broken.tm_mday;
+      start_time.tm_hour = time_broken.tm_hour;
+      start_time.tm_min = time_broken.tm_min;
 
       g_string_append_printf (xml, "<duration>%ld</duration>", duration);
     }
   else
     {
-      struct tm *time_broken;
-      time_broken = localtime (&now);
+      struct tm time_broken;
+      localtime_r (&now, &time_broken);
 
       start_year = params_value (params, "start_year");
       start_month = params_value (params, "start_month");
@@ -10013,21 +10024,20 @@ get_system_reports_gmp (gvm_connection_t *connection,
       end_minute = params_value (params, "end_minute");
 
       start_time.tm_year =
-        start_year ? atoi (start_year) - 1900 : time_broken->tm_year;
+        start_year ? atoi (start_year) - 1900 : time_broken.tm_year;
       start_time.tm_mon =
-        start_month ? atoi (start_month) - 1 : time_broken->tm_mon;
-      start_time.tm_mday = start_day ? atoi (start_day) : time_broken->tm_mday;
-      start_time.tm_hour =
-        start_hour ? atoi (start_hour) : time_broken->tm_hour;
+        start_month ? atoi (start_month) - 1 : time_broken.tm_mon;
+      start_time.tm_mday = start_day ? atoi (start_day) : time_broken.tm_mday;
+      start_time.tm_hour = start_hour ? atoi (start_hour) : time_broken.tm_hour;
       start_time.tm_min =
-        start_minute ? atoi (start_minute) : time_broken->tm_min;
+        start_minute ? atoi (start_minute) : time_broken.tm_min;
 
       end_time.tm_year =
-        end_year ? atoi (end_year) - 1900 : time_broken->tm_year;
-      end_time.tm_mon = end_month ? atoi (end_month) - 1 : time_broken->tm_mon;
-      end_time.tm_mday = end_day ? atoi (end_day) : time_broken->tm_mday;
-      end_time.tm_hour = end_hour ? atoi (end_hour) : time_broken->tm_hour;
-      end_time.tm_min = end_minute ? atoi (end_minute) : time_broken->tm_min;
+        end_year ? atoi (end_year) - 1900 : time_broken.tm_year;
+      end_time.tm_mon = end_month ? atoi (end_month) - 1 : time_broken.tm_mon;
+      end_time.tm_mday = end_day ? atoi (end_day) : time_broken.tm_mday;
+      end_time.tm_hour = end_hour ? atoi (end_hour) : time_broken.tm_hour;
+      end_time.tm_min = end_minute ? atoi (end_minute) : time_broken.tm_min;
     }
 
   g_string_append_printf (xml,
@@ -13407,7 +13417,7 @@ get_feeds_gmp (gvm_connection_t *connection, credentials_t *credentials,
   char *text = NULL;
   gchar *response;
   time_t now;
-  struct tm *tm;
+  struct tm tm;
   gchar current_timestamp[30];
 
   if (gvm_connection_sendf (connection, "<get_feeds/>") == -1)
@@ -13435,9 +13445,8 @@ get_feeds_gmp (gvm_connection_t *connection, credentials_t *credentials,
     }
 
   time (&now);
-  tm = gmtime (&now);
-  if (tm == NULL
-      || (strftime (current_timestamp, 29, "%Y-%m-%dT%H:%M:%S", tm) == 0))
+  if (gmtime_r (&now, &tm) == NULL
+      || (strftime (current_timestamp, 29, "%Y-%m-%dT%H:%M:%S", &tm) == 0))
     {
       current_timestamp[0] = '\0';
     }